Refactor : 장소 댓글 입력 파일 별도 분리
장소 댓글 입력 부분을 댓글 파일에서 별도 분리하여
코드의 가독성을 향상시킴
